Allow custom type detection on load (#128)
* Allow custom type detection on load

This adds get_data_type() as a way of specializing the way that
the type is detected during loading. This correlates with
`get_obj_type`, just with data moving in the opposite direction (load
vs dump).

This makes it possible to specify a OneOfSchema with more "bespoke"
ways of detecting the data's type than just checking a single string
field.

It's easy to specify a valid OneOf case in OpenAPI or similar where
there is no specific field which determines the type of the data. For
example, you might have different schemas depending on whether or not
a specific field is present at all -- this is the scenario used in the
testsuite for testing.

* Add changelog entry for #128

Co-authored-by: Steven Loria <sloria1@gmail.com>